Update GithubCard.vue
The user is fetched asynchronously; therefore, it takes some time. At first, it is undefined and only exists after the request to the API is complete; therefore everything will be displayed in the template after the request is done. consequently, first should be checked if the user exists; then all the markups after the user is fetched will be displayed (To avoid displaying incorrect or incomplete information, it checks if the user data is available before displaying content in the template).